requirements_cuda.txt
tensorflow
tensorflow-io
hydra_core==0.11.3
hydra_colorlog==0.1.4
pystoi==0.3.3
git+https://github.com/ludlows/python-pesq#egg=pesq
six